moving header into component and some cleanUp
