+++ code/frontend/src/pages/layout/Layout.module.css
@@ -6,189 +6,7 @@
     margin-bottom: -24px;
 }
 
-.header {
-    height: 56px;
-    background-color: rgb(0, 32, 91);
-    display: flex;
-}
-
-.headerContainer {
-    display: flex;
-    justify-content: left;
-    align-items: center;
-}
-
-.headerTitleContainer {
-    display: flex;
-    align-items: center;
-    margin-left: 30px;
-    margin-top: -5px;
-    text-decoration: none;
-    border-right: 1px solid #919CB0;
-    padding-right: 17px;
-    height: 24px;
-    width: 110px;
-}
-
-.headerTitle {
-    /* font-family: "Segoe UI"; */
-    /* font-style: normal; */
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 24px;
-    display: flex;
-    align-items: flex-end;
-    color: white;
-    margin-left: 35px;
-}
-
-.headerIcon {
-    /* height: 57px;
-    width: 100px; */
-    /* margin-left: 36px; */
-    width: 100%;;
-}
-
-.mainNavLinks {
-  position:absolute;
-  display: flex;
-  direction: row;
-  align-items: center;
-  right:15px;
-  color: #ffffff;
-  gap: 20px;
-}
-
-.shareButtonContainer {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    width: 32px;
-    height: 32px;
-    border-radius: 4px;
-    cursor: pointer;
-}
-
-.shareButton, .addSourceButton {
-    /* width: 20px; */
-    /* height: 20px; */
-    color: #FFFFFF;
-}
-
-.userMenuContainer {
-  display: flex;
-  flex-direction: row;
-}
-
-.userMenuBtn {
-  display: flex;
-  flex-direction: row;
-  margin: 4px 8px;
-  align-items: center;
-  gap: 8px;
-}
-
-.userMenuArrows {
-  font-weight: bold;
-  font-size: 8px;
-  transform: rotate(90deg) scale(1, 2);
-}
-
-.urlTextBox {
-    font-family: "Segoe UI";
-    font-style: normal;
-    font-weight: 400;
-    font-size: 14px;
-    line-height: 20px;
-    color: #707070;
-    border: 1px solid #D1D1D1;
-    border-radius: 4px;
-}
-
-.copyButtonContainer {
-    box-sizing: border-box;
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    padding: 5px 12px;
-    gap: 4px;
-    background: #FFFFFF;
-    border: 1px solid #D1D1D1;
-    border-radius: 4px;
-}
-
-.copyButtonContainer:hover {
-    cursor: pointer;
-    background: #D1D1D1;
-}
-
-.copyButton {
-    color: #424242;
-}
-
-.copyButtonText {
-    font-family: "Segoe UI";
-    font-style: normal;
-    font-weight: 600;
-    font-size: 14px;
-    line-height: 20px;
-    display: flex;
-    align-items: center;
-    color: #242424;
-}
-
-
 .disabled {
   cursor: not-allowed;
   opacity: 0.35;
-}
-
-
-.dataSourceToolTip {
-  background-color: #00205B !important;
-  color: #ffffff !important;
-  margin-top: 13px !important;
-  border-radius: 3px !important;
-}
-
-.dataSourceToolTipInner {
-  padding: 3px 2px;
-}
-
-.dataSourceToolTipInner span{
-  font-size: 14px !important;
-  font-weight: 400 !important;
-}
-
-.shareToolTip {
-  background-color: #00205B !important;
-  color: #ffffff !important;
-  margin-top: 10px !important;
-  border-radius: 3px !important;
-}
-
-.shareToolTipInner {
-  padding: 3px 2px;
-}
-
-.shareToolTipInner span{
-  font-size: 14px !important;
-  font-weight: 400 !important;
-}
-
-.copiedToClipboard{
-  display: flex;
-  align-items: center;
-  gap: 3px;
-}
-
-.copiedToClipboard img{
-
-}
-
-.copiedToClipboard span{
-  font-size: 14px !important;
-  font-weight: 400 !important;
 }
